refactor: replace PostCard structure with PostCardItem component
- Replaced individual PostCardHeader, PostCardMainContent, and PostCardFooter components with a unified PostCardItem component for improved modularity and readability.
- Streamlined the rendering of posts within the PostCard component, enhancing maintainability.
