+++ resources/js/components/post/comments/PostCommentsHeader.jsx
@@ -1,0 +1,9 @@
+import { getTitleRowClass, getTitleTextClass } from './commentSectionClassNames';
+
+export default function PostCommentsHeader({ isFacebookEmbed, embedded }) {
+    return (
+        <div className={getTitleRowClass(isFacebookEmbed, embedded)}>
+            <h2 className={getTitleTextClass(isFacebookEmbed, embedded)}>Comments</h2>
+        </div>
+    );
+}
